1/30
This is the first JS 30 exercise that I did, I watched the video and started doing the exercise on my own… I learned a couple of things like the kbd that I didn’t know and the data attributes. I have been studying web development for 2 months and I feel that I have made a lot of progress. Any contribution or comment on my codes, I will take it into account